Create README.md
Hello everyone! 👋
This is the repository of my website dedicated to games: https://github.com/
On the site, you will find a collection of different games that you can enjoy and train your skills!
How to access the website:You can visit the website directly at this address: nicolascassese.altervista.org

Instructions for playing:
Visit the website address mentioned above.
Explore the page and choose the game that interests you the most.
Click on the selected game and... have fun training!